pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synthadoc"
version = "0.1.0"
description = "Domain-agnostic LLM knowledge compilation engine"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "aiosqlite>=0.20",
    "anthropic>=0.30",
    "openai>=1.40",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pypdf>=4.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "rank_bm25>=0.2",
    "filelock>=3.15",
    "opentelemetry-sdk>=1.25",
    "pydantic>=2.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vectors = [
    "fastembed>=0.3",   # requires Rust toolchain to build
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "pytest-benchmark>=4.0",
    "respx>=0.21",
    "mcp>=1.0",
]

[project.scripts]
synthadoc = "synthadoc.cli.main:app"

[project.entry-points."synthadoc.skills"]
pdf      = "synthadoc.skills.pdf:PdfSkill"
url      = "synthadoc.skills.url:UrlSkill"
markdown = "synthadoc.skills.markdown:MarkdownSkill"
docx     = "synthadoc.skills.docx:DocxSkill"
xlsx     = "synthadoc.skills.xlsx:XlsxSkill"
image    = "synthadoc.skills.image:ImageSkill"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["synthadoc"]
